make PhpType for "mixed" be "\\var", not "var"
This fixes a bug in UI, that can't be expressed with unit tests.
Having an interface with @param mixed $arg, create an implementing class and click "add method stubs".
Earlier it failed with an assertion isFqn() inside reference extractor.
It still fails on some complex types like tuples, but I don't know the way to fix.