rename 'var' to 'mixed'/'kmixed'
Earlier, KPHP had the 'var' type, presented as 'mixed' in PHPDoc.
Later, 'var' was renamed to 'mixed' in KPHP internals.
Here we rename 'var' to 'mixed' in KPHPStorm.
Note, that 'mixed' from PHPDoc is expressed as 'kmixed' internally,
not to be messed up with auto-inferred 'mixed', which means 'any' in the plugin.